fix: reservation > query hook 위치 수정
